added schema sql script, always run on start. Resets DB to standard tables with some test data.
mapped user class to users table

endpoint /rest/Users implemented, returns all known users in JSON.

refactored sources structure
