+++ src/main/resources/application.properties
@@ -3,3 +3,4 @@ spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.password=password
 spring.datasource.username=username
 spring.datasource.url=jdbc:mysql://localhost:3306/databasename
+spring.sql.init.mode=always
